Added Project
Homework task to exercise MVC architecture and response to an http request. The controller class takes a variable "loopType" and directs the flow to one of three service classes. The response is based on an alphabetical array that gets built with one of three loop types (while, for, foreach). The result will be either a reverse array, a partial array up to a certain letter or a partial array with only every second letter.